phase2/backend/requirements.txt
# SalaatFlow Phase II - Backend Dependencies
# Python 3.11+

# FastAPI and server
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Database
sqlmodel==0.0.14
psycopg2-binary==2.9.9
alembic==1.13.1

# Environment and utilities
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS
fastapi-cors==0.0.6
